requirements.txt
Flask
mysql-connector-python
python-dotenv
pytesseract
opencv-python
numpy
Pillow
qrcode
twilio